+++ tasks/util/math.ts
@@ -1,0 +1,11 @@
+
+export function calculateOutputAmount(
+    amountIn: bigint,
+    reserveIn: bigint,
+    reserveOut: bigint,
+): bigint {
+    const amountInWithFee = amountIn * BigInt(997);
+    const numerator = amountInWithFee * reserveOut;
+    const denominator = reserveIn * BigInt(1000) + amountInWithFee;
+    return numerator / denominator;
+}
